8369828: Generalize share/utilities/bytes.hpp
